+++ Example/Source/Examples/Private/Renderer/Mesh/Mesh.cpp
@@ -257,7 +257,7 @@ void Mesh::onDraw()
 					renderTarget->getWidthAndHeight(width, height);
 
 					// Get the aspect ratio
-					aspectRatio = static_cast<float>(width) / height;
+					aspectRatio = static_cast<float>(width) / static_cast<float>(height);
 				}
 			}
